Fix bug #42
Analysis and fix supplied by JeffRo

This fix is primarily needed for the public version of Component Diagnostics.  The public version targets multiple versions of VS (Dev11 and higher), and as such uses the least common version of MPF, MS.VS.Shell.11.0.dll.  That version's RunningDocumentTable and RunningDocumentInfo classes had no knowledge of lazy tabs and would defeat it by calling GetDocumentInfo for every document, regardless of its state.  In addition to defeating lazy tab, it also resulted in misleading results in the RDT diagnostics window, such as those described in bug 915627.

The fix is to not rely on MPF's version of RunningDocumentTable and RunningDocumentInfo classes.  The relevant portions of lazy tab-friendly classes are copied into the Component Diagnostics package.

git-tfs-id: [https://paulharrington.visualstudio.com/defaultcollection]$/Component Diagnostics;C109